Open a configuration interface for default volumesnapshotclass (#10)
This app is provider-agnostic app and it shouldn't create
any provider specific default by default. That is why we did
the default volumesnapshotclass manifests in other apps. However,
it creates some dependency between apps and make the initialization
process slow&complex.

With this PR, we continue to not create any provider specific thing
by default. On the other hand, we will allow our consumers to create
default volumesnapshotclass with this app without introducing any inter-app
dependency.